Cream nat/pro filterjob (#15)
* update ver.1

* fix: refactor & fix register feat bugs

* fix: welcome back our job data

* fix: kill  some bugs

* fixed: filter find that job bug

* update ver.2

* refactor: backend code and fix some bugs

* fixed: can login with right role account

* fixed: extra bugs

* update: route path

* update job router

+++ server/utils/upload.js
@@ -1,19 +1,40 @@
 import { v2 as cloudinary } from "cloudinary";
 import fs from "fs/promises";
-const cloudinaryUpload = async (files) => {
+const cloudinaryUploadCV = async (files) => {
   const fileUrl = [];
-  for await (let file of files.uploadFile) {
-    const result = await cloudinary.uploader.upload(file.path, {
-      folder: "register/uploadfile",
-      type: "private",
-    });
-    fileUrl.push({
-      url: result.secure_url,
-      publicId: result.public_id,
-    });
-    await fs.unlink(file.path);
+  if (files.cvFile !== undefined) {
+    for (let file of files.cvFile) {
+      const result = await cloudinary.uploader.upload(file.path, {
+        folder: "register/cvFile",
+        type: "private",
+      });
+      fileUrl.push({
+        url: result.secure_url,
+        publicId: result.public_id,
+      });
+      await fs.unlink(file.path);
+    }
+    return fileUrl;
   }
-
-  return fileUrl;
 };
-export { cloudinaryUpload };
+
+const cloudinaryUploadLogo = async (files) => {
+  const logoUrl = [];
+  if (files.logoFile !== undefined) {
+    for (let file of files.logoFile) {
+      const result = await cloudinary.uploader.upload(file.path, {
+        folder: "register/logoFile",
+        type: "private",
+      });
+      logoUrl.push({
+        url: result.secure_url,
+        publicId: result.public_id,
+      });
+      await fs.unlink(file.path);
+    }
+
+    return logoUrl;
+  }
+};
+
+export { cloudinaryUploadCV, cloudinaryUploadLogo };
